+++ spec/fabricators/match_fabricator.rb
@@ -2,11 +2,6 @@ Fabricator(:match_lost_to, class_name: 'Match') do
   after_build do |match|
     match.scores = match.tied? ? [[3, 3]] : [[15, 21]]
   end
-  after_create do |match|
-    match.winners.inc(wins: 1)
-    match.losers.inc(losses: 1)
-    User.rank!(match.team)
-  end
 end
 
 Fabricator(:match) do
@@ -18,9 +13,6 @@ Fabricator(:match) do
     match.scores = match.tied? ? [[3, 3]] : [[15, 21]]
   end
   after_create do |match|
-    match.winners.inc(wins: 1)
-    match.losers.inc(losses: 1)
-    User.rank!(match.team)
     challenge.update_attributes!(state: ChallengeState::PLAYED, updated_by: match.losers.first)
   end
 end